Fixups.

+++ SerialPrograms/Source/PokemonSwSh/Programs/NonShinyHunting/PokemonSwSh_StatsReset-Calyrex.cpp
@@ -216,12 +216,12 @@ void StatsResetCalyrex::program(SingleSwitchProgramEnvironment& env, BotBaseCont
         for (int i = 0; i < 40; i++){
             pbf_press_button(context, BUTTON_A, 10, 1 * TICKS_PER_SECOND);
         }
-        pbf_press_button(context, BUTTON_X  , 10, 1.5 * TICKS_PER_SECOND);
-        pbf_press_dpad  (context, DPAD_RIGHT, 10, 0.5 * TICKS_PER_SECOND);
-        pbf_press_dpad  (context, DPAD_RIGHT, 10, 0.5 * TICKS_PER_SECOND);
+        pbf_press_button(context, BUTTON_X  , 10, (uint16_t)(1.5 * TICKS_PER_SECOND));
+        pbf_press_dpad  (context, DPAD_RIGHT, 10, (uint16_t)(0.5 * TICKS_PER_SECOND));
+        pbf_press_dpad  (context, DPAD_RIGHT, 10, (uint16_t)(0.5 * TICKS_PER_SECOND));
         pbf_press_button(context, BUTTON_A  , 10, 2   * TICKS_PER_SECOND);
-        pbf_press_dpad  (context, DPAD_LEFT , 10, 0.5 * TICKS_PER_SECOND);
-        pbf_press_dpad  (context, DPAD_UP   , 10, 0.5 * TICKS_PER_SECOND);
+        pbf_press_dpad  (context, DPAD_LEFT , 10, (uint16_t)(0.5 * TICKS_PER_SECOND));
+        pbf_press_dpad  (context, DPAD_UP   , 10, (uint16_t)(0.5 * TICKS_PER_SECOND));
         pbf_press_button(context, BUTTON_A  , 10, 2   * TICKS_PER_SECOND);
         pbf_press_button(context, BUTTON_A  , 10, 2   * TICKS_PER_SECOND);
         pbf_press_dpad  (context, DPAD_UP   , 10, 2   * TICKS_PER_SECOND);
@@ -234,7 +234,7 @@ void StatsResetCalyrex::program(SingleSwitchProgramEnvironment& env, BotBaseCont
             pbf_press_button(context, BUTTON_B, 10, 1 * TICKS_PER_SECOND);
         }
         pbf_press_button(context, BUTTON_X , 10, 2   * TICKS_PER_SECOND);
-        pbf_press_dpad  (context, DPAD_LEFT, 10, 0.5 * TICKS_PER_SECOND);
+        pbf_press_dpad  (context, DPAD_LEFT, 10, (uint16_t)(0.5 * TICKS_PER_SECOND));
         pbf_press_button(context, BUTTON_A , 10, 2   * TICKS_PER_SECOND);
         pbf_press_button(context, BUTTON_R , 10, 3   * TICKS_PER_SECOND);
         pbf_press_dpad  (context, DPAD_LEFT, 10, 1   * TICKS_PER_SECOND);
